little_shark_binance_v_1/requirements.txt
matplotlib==3.6.3
numpy==1.24.1
pandas==1.5.2
protobuf==4.23.4
python_binance==1.0.15
remote_logger==0.3.1
scipy==1.11.1
sentry_sdk==1.28.1
statsmodels==0.13.5
